Description, developer tools and other | v. 2.0a
This update adds developer tools for ease of working with databases, adding new texts, db strings and other.
Now mastervault operates on a sqlite database which can be replaced depending on user language.

Not all tools are yet developed, there are be bugs.

This is quite an old project of mine, Now that I rediscovered it I will try to develop it further to add the necessary functionalities of a mastervault.

Version 2.0 Alpha