Text operations and new database rules
Fixed some try statements so it rolls back changing enum if there is a database error
Added some more checks in databases.py add.

New rule
Databases should now not be distributed with the program, as it leaves potential for vulnerabilities and is hard to merge if conflicts exist. Databases should be dumped as sql and should be created from a sql

Version 2.0.2 Alpha